+++ src/Wrapper/index.types.ts
@@ -1,5 +1,6 @@
 // Types
 import type { BoxProps } from '@mui/material'
+import type { ResponsiveStyleValue, GridSpacing } from '@mui/system'
 
 export type SpinoramaWrapperProps = BoxProps & {
 	/**
@@ -19,4 +20,11 @@ export type SpinoramaWrapperProps = BoxProps & {
 	 * @default false
 	 */
 	reverse?: boolean
+
+	/**
+	 * Defines the space between the type `item` components.
+	 * It can only be used on a type `container` component.
+	 * @default 0
+	 */
+	spacing?: ResponsiveStyleValue<GridSpacing>
 }