+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Release Notes
 All notable changes to this package will be documented in this file.
 
+- Added the `spacing` prop to the **SpinoramaThumbnails** and **SpinoramaWrapper** components
+
 # 0.4.26
 - Added the ability to provide custom buttons to the **SpinoramaThumbnail** component